creato il js per nascondere il bottone "scegli un file" del form inserisci prodotto
creata la servlet che controlla se l'utente che accede a determinate
pagine della parte admin iano degli admin

creato il js per la validazione dei prodotti con le funzioni regex

creato  il  js per l'utilizzo delle funzioni js su ogni campo del form
dell'inseirmneto dei prodotti

vari aggiustature per far combaciare tutto

+++ WebProgramming/src/main/java/control/RegistraOrdinazioneServlet.java
@@ -80,9 +80,9 @@ public class RegistraOrdinazioneServlet extends HttpServlet {
 		if(statoOrdinazione==null || statoOrdinazione.trim().isEmpty()) {
 			errori.add("Il campo stato non può essere vuoto");
 		}
-		float importo=0;
+		double importo=0;
 		try {
-			importo= Float.parseFloat(importoStr);
+			importo= Double.parseDouble(importoStr);
 			if(importo<0) {
 				errori.add("L'importo non può essere negativo");
 			}
